Bugfix overwrite correct templates when using multiphp
- multiphp upgrade script used a undefine variable

- When rebuilding web domains, do not assume deprecated config file to be present

+++ func/domain.sh
@@ -307,10 +307,12 @@ del_web_config() {
     fi
 
     # Remove domain configuration files and clean up symbolic links
-    rm -f /etc/$WEB_SYSTEM/conf.d/domains/$domain.conf 
-    rm -f /etc/$PROXY_SYSTEM/conf.d/domains/$domain.conf 
-    rm -f /etc/$WEB_SYSTEM/conf.d/domains/$domain.ssl.conf 
-    rm -f /etc/$PROXY_SYSTEM/conf.d/domains/$domain.ssl.conf 
+    rm -f /etc/$WEB_SYSTEM/conf.d/domains/$domain.conf
+    rm -f /etc/$WEB_SYSTEM/conf.d/domains/$domain.ssl.conf
+    if [ ! -z $PROXY_SYSTEM ]; then
+        rm -f /etc/$PROXY_SYSTEM/conf.d/domains/$domain.conf 
+        rm -f /etc/$PROXY_SYSTEM/conf.d/domains/$domain.ssl.conf 
+    fi
 
     # Clean up legacy configuration files
     if [ -e "$conf" ]; then
